pyproject.toml
[project]
name = "der-duemmste"
version = "0.1.0"
description = "Game Interface to manage questions and create a Browser Overlay for OBS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eventlet>=0.39.0",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]

[project.scripts]
der-duemmste = "der_duemmste:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uvicorn]
host = "0.0.0.0"
port = 5000
workers = 2
log_level = "info"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]

